Shellcheck fixes, conditional file deletions for prep for new id
Various shellcheck fixes.
Make files/folders for prepare-for-new-identity.sh conditional.
- DeviceDB for example is not included in the latest releases, so that
  file is with very high likelyhood not there at all.

+++ example-scripts/meta-pelion-edge/prepare-for-new-identity.sh
@@ -1,4 +1,8 @@
+#!/bin/bash
+#
 # Copyright (c) 2020, Arm Limited and affiliates.
+# Copyright (c) 2023, Izuma Networks
+#
 # SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -26,26 +30,35 @@ systemctl stop edge-core
 echo "Stopped edge-core"
 
 # Remove old device management credentials
-rm -rf $PDM_CRED_DIR
-echo "Cleared device management credentials"
+rm -rf "$PDM_CRED_DIR"
+echo "Cleared device management credentials ($PDM_CRED_DIR -directory)"
 
 # Remove old identity.json
-rm -rf $EDGE_GW_CONFIG_DIR
-echo "Cleared identity.json"
+rm -rf "$EDGE_GW_CONFIG_DIR"
+echo "Cleared identity.json ($EDGE_GW_CONFIG_DIR -directory)"
 
 # -------------------------------------------------------------------
 # meta-pelion-edge platform specifics
 # -------------------------------------------------------------------
 # Remove the gateway statistics generated from old provisioning data. This is used by info command
-rm -rf /wigwag/system/lib/bash/relaystatics.sh
-echo "Cleared relaystatics.sh"
+if [ -e "/wigwag/system/lib/bash/relaystatics.sh" ]; then
+    rm -rf /wigwag/system/lib/bash/relaystatics.sh
+    echo "Cleared relaystatics.sh"
+fi
 
-# Remove device database
-rm -rf /userdata/etc/devicejs/db
-echo "Cleared devicedb"
+# Remove device database, be it a file or folder
+# Edge 2.5.0/2.6.0 does not have this.
+DEVICEDB="/userdata/etc/devicejs/db"
+if [ -e "$DEVICEDB" ] || [ -d "$DEVICEDB" ]; then
+    rm -rf "$DEVICEDB"
+    echo "Cleared devicedb ($DEVICEDB -folder/file)"
+fi
 
-# Remove maestro database
-rm -rf /userdata/etc/maestroConfig.db
-echo "Cleared maestro database"
+# Remove maestro database, be it a file or folder
+MAESTRO="/userdata/etc/maestroConfig.db"
+if [ -e "$MAESTRO" ] || [ -d "$MAESTRO" ]; then
+    rm -rf "$MAESTRO"
+    echo "Cleared maestro database ($MAESTRO -directory)"
+fi
 
 echo "Done"
